# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DocFlow AI Document Processing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/{id}/validations:
    get:
      operationId: getValidations
      tags: [validations]
      summary: Get validation issues for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validation issues list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationIssue"

  # ── Review ─────────────────────────────────────────────────────────────────
...
  /workflow-tasks:
    get:
      operationId: listWorkflowTasks
      tags: [workflows]
      summary: List workflow approval tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, escalated]
        - name: assignedTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Workflow tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowTask"
    post:
      operationId: createWorkflowTask
      tags: [workflows]
      summary: Create a workflow task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTask"
  /workflow-tasks/{id}/action:
    post:
      operationId: actOnWorkflowTask
      tags: [workflows]
      summary: Approve, reject or escalate a workflow task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowActionInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTask"

  # ── Vendors ─────────────────────────────────────────────────────────────────
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Vendors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
    patch:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorUpdate"
      responses:
        "200":
          description: Updated vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

  # ── Exports ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ValidationIssue:
      type: object
      required: [id, documentId, ruleCode, severity, message, status]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        ruleCode:
          type: string
        severity:
          type: string
          enum: [error, warning, info]
        message:
          type: string
        field:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, resolved, ignored]

    # ── Review ──────────────────────────────────────────────────────────────────
...
    WorkflowTask:
      type: object
      required: [id, documentId, taskType, status, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        taskType:
          type: string
          enum: [review, approval, export, escalation]
        status:
          type: string
          enum: [pending, approved, rejected, escalated]
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        comment:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        completedBy:
          type: ["string", "null"]
        createdAt:
          type: string
    WorkflowTaskInput:
      type: object
      required: [documentId, taskType]
      properties:
        documentId:
          type: integer
        taskType:
          type: string
          enum: [review, approval, export, escalation]
        assignedTo:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
    WorkflowActionInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject, escalate]
        comment:
          type: ["string", "null"]
        completedBy:
          type: string

    # ── Vendor ──────────────────────────────────────────────────────────────────
    Vendor:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        gstin:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        isActive:
          type: boolean
        totalDocuments:
          type: integer
        createdAt:
          type: string
    VendorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        gstin:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
    VendorUpdate:
      type: object
      properties:
        name:
          type: string
        gstin:
          type: ["string", "null"]
        pan:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── Export ──────────────────────────────────────────────────────────────────
...